requirements.txt
pyyaml
uuid
